feat: 소극장 공연 등록 구현 및 수정
