+++ .gitignore
@@ -36,4 +36,6 @@ out/
 ### VS Code ###
 .vscode/
 
-**/application.yml
+!**src/main/resources/application.yml
+!**src/main/resources/application-dev.yml
+!**src/main/resources/application-prod.yml
