+++ src/main/java/cc/backend/amateurShow/entity/AmateurCasting.java
@@ -1,26 +1,31 @@
-package cc.backend.domain.entity.amateur;
+package cc.backend.amateurShow.entity;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import cc.backend.domain.common.BaseEntity;
 
 @Entity
 @Getter
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class AmateurStaff {
+public class AmateurCasting extends BaseEntity {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private String position;
+    private String imageUrl;
 
-    private String name;
+    private String actorName;
+
+    private String castingName;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "amateur_show_id")
     private AmateurShow amateurShow;
+
 }
