+++ src/main/java/cc/backend/amateurShow/entity/AmateurNotice.java
@@ -1,0 +1,31 @@
+package cc.backend.amateurShow.entity;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import cc.backend.domain.common.BaseEntity;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Entity
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class AmateurNotice extends BaseEntity {
+
+    @Id@GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ElementCollection
+    private List<String> noticeImageUrls = new ArrayList<>();
+
+    private String content;
+
+    @OneToOne(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
+    @JoinColumn(name = "amateur_show_id")
+    private AmateurShow amateurShow;
+}
