+++ src/main/java/cc/backend/amateurShow/entity/AmateurStaff.java
@@ -1,31 +1,26 @@
-package cc.backend.domain.entity.amateur;
+package cc.backend.amateurShow.entity;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import cc.backend.domain.common.BaseEntity;
 
 @Entity
 @Getter
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class AmateurCasting extends BaseEntity {
-
+public class AmateurStaff {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private String imageUrl;
+    private String position;
 
-    private String actorName;
-
-    private String castingName;
+    private String name;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "amateur_show_id")
     private AmateurShow amateurShow;
-
 }
