+++ src/main/java/cc/backend/amateurShow/entity/enums/TicketType.java
@@ -1,0 +1,5 @@
+package cc.backend.amateurShow.entity.enums;
+
+public enum TicketType {
+    COMMON, DISCOUNT
+}
