+++ src/main/java/cc/backend/amateurShow/repository/AmateurCastingRepository.java
@@ -1,6 +1,6 @@
-package cc.backend.repository.amateurRepository;
+package cc.backend.amateurShow.repository;
 
-import cc.backend.domain.entity.amateur.AmateurCasting;
+import cc.backend.amateurShow.entity.AmateurCasting;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
