+++ src/main/java/cc/backend/amateurShow/repository/AmateurShowRepository.java
@@ -1,0 +1,12 @@
+package cc.backend.amateurShow.repository;
+
+import cc.backend.amateurShow.entity.AmateurShow;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+@Repository
+public interface AmateurShowRepository extends JpaRepository<AmateurShow, Long> {
+    List<AmateurShow> findAllByMemberId(Long memberId);
+
+}
