--- src/main/java/cc/backend/domain/entity/amateur/AmateurSummary.java
@@ -1,33 +1,0 @@
-package cc.backend.domain.entity.amateur;
-
-import cc.backend.domain.common.BaseEntity;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class AmateurSummary extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String content;
-
-    private String summaryImage;
-
-    @OneToOne(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
-    @JoinColumn(name = "amateur_show_id")
-    private AmateurShow amateurShow;
-
-    // amateurShow 업데이트할 때 사용
-    public void updateContent(String newContent) {
-        this.content = newContent;
-    }
-}
